a-2025-job/i_youlify/fastapi_billing_demo/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
